Coding standards revision part 2.
More violations fixed. Quick test shows app appears to function
properly. There are 4 remaining violations that deal with using
system.out.print in sections that deal with exception handling for
CurrentUser and UserList. I have left these as a bookmark in case we
want to do something else in case of an exception instead of me just
deleting the line so nothing happens. Still have not managed to get
findbugs working.